started writing presenters specs, ran into a viewtastic test-env activation issue

+++ app/presenters/user_presenter.rb
@@ -1,9 +1,6 @@
 class UserPresenter < Viewtastic::Base
-  presents :user
-  
-  delegate :id, :login, :email, :active?, :has_no_credentials?, :avatar,
-           :to => :user
-           
+  presents :user => [:id, :login, :email, :active?, :has_no_credentials?, :avatar]
+             
   delegate :current_user, :current_team,
           :to => :controller
   
@@ -14,8 +11,8 @@ class UserPresenter < Viewtastic::Base
   def action_links
     returning([]) do |links|
       if current_user.account_holder?
-        links << link_to("Remove", team_user_path(user), :method => :delete, :class => "delete", :confirm => "Are you sure you want to remove #{user}?")
-        links << link_to(user.active? ? 'Disable' : 'Enable', [:alter_status, :team, user], :method => :post, :class => user.active? && "delete")
+        links << link_to("Remove", team_user_path(user), :method => :delete, :class => "delete", :confirm => "Are you sure you want to remove #{user}?") unless current_user == user
+        links << link_to(user.active? ? 'Disable' : 'Enable', [:alter_status, :team, user], :method => :post, :class => user.active? && "delete") unless current_user == user
       end
     end    
   end
